Add unit test for aggregation and fix flappy tests
Added big unit test to verify new Aggregate() functionality based on
curren state in geckon.

Due to use of map in couple function implementations, order of function
result was not deterministic and since there was no sorting in unit
tests, the test results were flappy.

It's useful to have deterministic results on functions so sorting was
added to actual implementations of business logic instead of unit tests.