Deduplicate versions in release bundles (#39)
When Aggregate() builds version bundle aggregates, it creates all
possible permutations allowed within dependency definitions.
NewRelease() then creates a release from a version bundle and as a part
of this process, release gets calculated version number that is sum of
bundle versions. Given suitable combination of different versions for
different version bundles there is high probability to get same release
version for different contents. This is resolved by selecting the
release with latest timestamp with the exception that active release
must not be dismissed.