Don't require non-empty component list for version bundle (#40)
There are situations where it's not useful to list any components for a
version bundle. Cluster-operator is one example here as theoretically
only reasonable component it could have is provider specific operator
such as aws-operator but it doesn't manage the version of it so listing
it in components would present misleading information in release bundle
component listing.