Remove release aggregation
Due to various issues in dynamic release aggregation, a new release
management implementation was introduced with IndexRelease. Therefore
old (and in many ways buggy) release aggregation shall be removed.

Some problems that dynamic release aggregation had:
- Evolution of version bundles cause duplicated release versions.
- Evolution of version bundles cause changes to release version
  contents.

Those issues then cause a lot of problems for running clusters as there
is now definitive answer to "what was predecessor of a specific release
and what would be the correct successor of it" especially when contents
of some specific release version could change due to version bundle
evolution.

Therefore this code should be retired and deleted.